refactor(linter/unicorn): extract optional chain logic
