+++ crates/oxc_linter/src/utils/unicorn.rs
@@ -20,6 +20,9 @@ use crate::LintContext;
 mod boolean;
 pub use boolean::*;
 
+mod optional_chain;
+pub use optional_chain::*;
+
 // Built-in Error constructors
 // https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Error#Error_types
 pub const BUILT_IN_ERRORS: [&str; 10] = [
